+++ src/lib/src/core/models/page-center-point.ts
@@ -1,18 +1,30 @@
-import { ArrayUtils } from './../viewer-service/array-utils';
 import { Point } from './point';
 
 export class CenterPoints {
   private centerPoints: Point[] = [];
 
-  public findClosestIndex(point: Point): number {
-    return new ArrayUtils().findClosestIndex(this.centerPoints, point.x);
-  }
-
-  public add(point: Point): void  {
+  public add(point: Point): void {
     this.centerPoints.push(point);
   }
 
   public get(index: number): Point {
-    return this.centerPoints[index];
+    return {...this.centerPoints[index]};
+  }
+
+  public findClosestIndex(point: Point): number {
+    let i: number;
+    let result: any;
+    let lastDelta: any;
+
+    this.centerPoints.some(function (item: Point, index: number) {
+      const delta = Math.abs(point.x - item.x);
+      if (delta >= lastDelta) {
+        return true;
+      }
+      i = index;
+      result = item;
+      lastDelta = delta;
+    });
+    return i;
   }
 }
