+++ src/lib/src/viewer/viewer.component.spec.ts
@@ -70,13 +70,13 @@ describe('ViewerComponent', function () {
 
   it('should create component', () => expect(comp).toBeDefined());
 
-  it('should close all dialogs when manifestUri changes', () => {
+  it('should cleanUp when manifestUri changes', () => {
     testHostComponent.manifestUri = 'dummyURI2';
 
-    spyOn(testHostComponent.viewerComponent.dialog, 'closeAll').and.callThrough();
+    spyOn(testHostComponent.viewerComponent, 'cleanUp').and.callThrough();
     testHostFixture.detectChanges();
 
-    expect(testHostComponent.viewerComponent.dialog.closeAll).toHaveBeenCalled();
+    expect(testHostComponent.viewerComponent.cleanUp).toHaveBeenCalled();
   });
 
   it('should create viewer', inject([ViewerService], (viewerService: ViewerService) => {
